requirements.txt
aenum==3.1.15
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
coverage==7.4.0
cryptography==41.0.1
dbf==0.99.9
exceptiongroup==1.2.0
flake8==7.0.0
Flask==3.0.1
fpdf==1.7.2
freezegun==1.4.0
future==0.18.3
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyparsing==3.1.1
PySimpleSOAP==1.8.22
pytest==7.2.0
pytest-cov==4.0.0
pytest-freezegun==0.4.2
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-mock==3.10.0
pytest-vcr==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==6.1
six==1.16.0
tabulate==0.8.5
tomli==2.0.1
vcrpy==5.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4